Fix: A byte in Java represents a signed 8 bit integer. (In C sharp a byte represents an unsigned 8 bit integer.)

+++ src/core/MathUtils.java
@@ -12,7 +12,7 @@ public class MathUtils {
         return (x < high) ? ((x > low) ? x : low) : high;
     }
 
-    public static byte toByte(double x, double gamma) {
-        return (byte)clamp(255.0 * Math.pow(x, 1 / gamma), 0.0, 255.0);
+    public static int toByte(double x, double gamma) {
+        return (int)clamp(255.0 * Math.pow(x, 1 / gamma), 0.0, 255.0);
     }
 }
